Refactor k-empty
